Create logout hook
